+++ services/backend/warehouse-service/src/main/java/com/veteroch4k/warehouse/models/commands/RequiredMaterial.java
@@ -1,6 +1,6 @@
 package com.veteroch4k.warehouse.models.commands;
 
-public record RequiredMaterial(int materialId,
-                               int quantity) {
+public record RequiredMaterial(Long materialId,
+                               Long quantity) {
 
 }
