+++ services/backend/warehouse-service/src/main/java/com/veteroch4k/warehouse/repositories/MaterialAccountingRepository.java
@@ -3,7 +3,7 @@ package com.veteroch4k.warehouse.repositories;
 import com.veteroch4k.warehouse.models.MaterialAccounting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialAccountingRepository extends JpaRepository<MaterialAccounting, Integer> {
+public interface MaterialAccountingRepository extends JpaRepository<MaterialAccounting, Long> {
 
 
 }
